sending users to new dashboard while keeping the old version under -old
